bugfix redis collector create too many instance case oom (#734)
* bugfix redis collector create too many instance

* bugfix redis collector create too many instance

* bugfix redis collector create too many instance
