+++ manager/src/main/resources/application.yml
@@ -101,7 +101,7 @@ warehouse:
 # 存储历史数据方式, 下方只能enabled启用一种方式
     jpa:
       enabled: true
-      expire-time: 7D
+      expire-time: 1h
     td-engine:
       enabled: false
       driver-class-name: com.taosdata.jdbc.rs.RestfulDriver
